pyproject.toml
[tool.poetry]
name = "qortex"
version = "0.1.0"
description = ""
authors = ["Nikita Kolesnik <118283668+KolesnikNV@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
djangorestframework = "^3.14.0"
loguru = "^0.7.0"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
faker = "^19.3.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
